+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 
 **Tools**
 
+* Added the Project Browser
 * Fixed an issue that caused the Level Editor to not include "core/units/camera" in the boot.package of a newly created project
 
 0.37.0